* t_ser.c: Cast getpid() calls to int as arguments to sprintf
	* ser_actx.c: Move prototypes (listed below) to int-proto.h

	* int-proto.h: Add prototypes for krb5_ser_authdata_init,
	krb5_ser_address_init, krb5_ser_authenticator_init,
	krb5_ser_checksum_init, krb5_ser_keyblock_init,
	krb5_ser_principal_init.

	* ser_adata.c, ser_addr.c, ser_auth.c, ser_cksum.c, ser_key.c,
	ser_princ.c: Include int-proto.h for prototypes.

By moving the prototypes to int-proto.h, it ndles the gcc warning
on missing prototypes. These functions are intenal and all used by
krb5_ser_auth_context_init()

git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@12814 dc483132-0cff-0310-8789-dd5450dbe970

+++ src/lib/krb5/krb/t_ser.c
@@ -212,7 +212,7 @@ ser_kcontext_test(kcontext, verbose)
     profile_t		sprofile;
     char		dbname[128];
 
-    sprintf(dbname, "temp_%d", getpid());
+    sprintf(dbname, "temp_%d", (int) getpid());
     sprofile = kcontext->profile;
     kcontext->profile = (profile_t) NULL;
     if (!(kret = ser_data(verbose, "> Context with no profile",
@@ -343,7 +343,7 @@ ser_acontext_test(kcontext, verbose)
 		 */
 		memset(&aent, 0, sizeof(aent));
 		aent.magic = KV5M_AUTHENTICATOR;
-		sprintf(clname, "help/me/%d@this.is.a.test", getpid());
+		sprintf(clname, "help/me/%d@this.is.a.test", (int) getpid());
 		actx->authentp = &aent;
 		if (!(kret = krb5_parse_name(kcontext, clname,
 					     &aent.client)) &&
@@ -394,9 +394,9 @@ ser_ccache_test(kcontext, verbose)
     krb5_principal	principal;
     extern krb5_cc_ops	krb5_scc_ops;
 
-    sprintf(ccname, "temp_cc_%d", getpid());
+    sprintf(ccname, "temp_cc_%d", (int) getpid());
     sprintf(princname, "zowie%d/instance%d@this.is.a.test",
-	    getpid(), getpid());
+	    (int) getpid(), (int) getpid());
     if (!(kret = krb5_cc_resolve(kcontext, ccname, &ccache)) &&
 	!(kret = ser_data(verbose, "> Resolved default ccache",
 			  (krb5_pointer) ccache, KV5M_CCACHE)) &&
@@ -406,9 +406,9 @@ ser_ccache_test(kcontext, verbose)
 			  (krb5_pointer) ccache, KV5M_CCACHE)) &&
 	!(kret = krb5_cc_destroy(kcontext, ccache))) {
 	krb5_free_principal(kcontext, principal);
-	sprintf(ccname, "FILE:temp_cc_%d", getpid());
+	sprintf(ccname, "FILE:temp_cc_%d", (int) getpid());
 	sprintf(princname, "xxx%d/i%d@this.is.a.test",
-		getpid(), getpid());
+		(int) getpid(), (int) getpid());
 	if (!(kret = krb5_cc_resolve(kcontext, ccname, &ccache)) &&
 	    !(kret = ser_data(verbose, "> Resolved FILE ccache",
 			      (krb5_pointer) ccache, KV5M_CCACHE)) &&
@@ -418,9 +418,9 @@ ser_ccache_test(kcontext, verbose)
 			      (krb5_pointer) ccache, KV5M_CCACHE)) &&
 	    !(kret = krb5_cc_destroy(kcontext, ccache))) {
 	    krb5_free_principal(kcontext, principal);
-	    sprintf(ccname, "STDIO:temp_cc_%d", getpid());
+	    sprintf(ccname, "STDIO:temp_cc_%d", (int) getpid());
 	    sprintf(princname, "xxx%d/i%d@this.is.a.test",
-		    getpid(), getpid());
+		    (int) getpid(), (int) getpid());
 	    if ((kret = krb5_cc_resolve(kcontext, ccname, &ccache)))
 		kret = krb5_cc_register(kcontext, &krb5_scc_ops, 1);
 	    if (!kret &&
@@ -456,17 +456,17 @@ ser_keytab_test(kcontext, verbose)
     krb5_keytab		keytab;
     extern krb5_kt_ops	krb5_ktf_writable_ops;
 
-    sprintf(ccname, "temp_kt_%d", getpid());
+    sprintf(ccname, "temp_kt_%d", (int) getpid());
     if (!(kret = krb5_kt_resolve(kcontext, ccname, &keytab)) &&
 	!(kret = ser_data(verbose, "> Resolved default keytab",
 			  (krb5_pointer) keytab, KV5M_KEYTAB)) &&
 	!(kret = krb5_kt_close(kcontext, keytab))) {
-	sprintf(ccname, "FILE:temp_kt_%d", getpid());
+	sprintf(ccname, "FILE:temp_kt_%d", (int) getpid());
 	if (!(kret = krb5_kt_resolve(kcontext, ccname, &keytab)) &&
 	    !(kret = ser_data(verbose, "> Resolved FILE keytab",
 			      (krb5_pointer) keytab, KV5M_KEYTAB)) &&
 	    !(kret = krb5_kt_close(kcontext, keytab))) {
-	    sprintf(ccname, "WRFILE:temp_kt_%d", getpid());
+	    sprintf(ccname, "WRFILE:temp_kt_%d", (int) getpid());
 	    if ((kret = krb5_kt_resolve(kcontext, ccname, &keytab)))
 		kret = krb5_kt_register(kcontext, &krb5_ktf_writable_ops);
 	    if (!kret &&
@@ -496,7 +496,7 @@ ser_rcache_test(kcontext, verbose)
     char		rcname[128];
     krb5_rcache		rcache;
 
-    sprintf(rcname, "dfl:temp_rc_%d", getpid());
+    sprintf(rcname, "dfl:temp_rc_%d", (int) getpid());
     if (!(kret = krb5_rc_resolve_full(kcontext, &rcache, rcname)) &&
 	!(kret = ser_data(verbose, "> Resolved FILE rcache",
 			  (krb5_pointer) rcache, KV5M_RCACHE)) &&
@@ -578,7 +578,7 @@ ser_princ_test(kcontext, verbose)
     krb5_principal	princ;
     char		pname[1024];
 
-    sprintf(pname, "the/quick/brown/fox/jumped/over/the/lazy/dog/%d@this.is.a.test", getpid());
+    sprintf(pname, "the/quick/brown/fox/jumped/over/the/lazy/dog/%d@this.is.a.test", (int) getpid());
     if (!(kret = krb5_parse_name(kcontext, pname, &princ))) {
 	if (!(kret = ser_data(verbose, "> Principal",
 			      (krb5_pointer) princ, KV5M_PRINCIPAL))) {
